Type mismatch fixes.

+++ zed_net.h
@@ -1,21 +1,20 @@
 /////////////////////////////////////////////////////////////////////////////////////////
 //
-// zed_net - v0.20 - public domain networking library
+// zed_net - v0.21 - public domain networking library
 // (inspired by the excellent stb libraries: https://github.com/nothings/stb)
 //
 // This library is intended primarily for use in games and provides a simple wrapper
 // around BSD sockets (Winsock 2.2 on Windows). Sockets can be set to be blocking or
 // non-blocking.
 //
-// Only UDP sockets are supported at this time, but this may later expand to include TCP.
-//
 // VERSION HISTORY
 //
-//    0.20 (7/28/2019) OSX compilation fixes.
-//    0.19 (3/4/2016) TCP added and malloc/free calls removed.
-//                     Not backwards compatible. - Ian T. Jacobsen (itjac.me)
-//    0.18 (9/13/2015) minor polishing
-//    0.17 (8/8/2015) initial release
+//    0.21 (12/24/2019) Type mismatch fixes.
+//    0.20 (7/28/2019)  OSX compilation fixes.
+//    0.19 (3/4/2016)   TCP added and malloc/free calls removed.
+//                      Not backwards compatible. - Ian T. Jacobsen (itjac.me)
+//    0.18 (9/13/2015)  Minor polishing
+//    0.17 (8/8/2015)   Initial release
 //
 // LICENSE
 //
@@ -97,7 +96,7 @@ ZED_NET_DEF const char *zed_net_host_to_str(unsigned int host);
 
 // Wraps the system handle for a UDP/TCP socket
 typedef struct {
-    int handle;
+    unsigned int handle;
     unsigned long non_blocking;
     int ready;
 } zed_net_socket_t;
@@ -117,7 +116,7 @@ ZED_NET_DEF void zed_net_socket_close(zed_net_socket_t *socket);
 //
 // Returns 0 on success
 // Returns -1 on failure (call 'zed_net_get_error' for more info)
-ZED_NET_DEF int zed_net_udp_socket_open(zed_net_socket_t *socket, unsigned int port, unsigned long non_blocking);
+ZED_NET_DEF int zed_net_udp_socket_open(zed_net_socket_t *socket, unsigned short port, unsigned long non_blocking);
 
 // Sends a specific amount of data to 'destination'
 //
@@ -143,7 +142,7 @@ ZED_NET_DEF int zed_net_udp_socket_receive(zed_net_socket_t *socket, zed_net_add
 // Socket will listen for incoming connections if 'listen_socket' is non-zero
 // Returns 0 on success
 // Returns -1 on failure (call 'zed_net_get_error' for more info)
-ZED_NET_DEF int zed_net_tcp_socket_open(zed_net_socket_t *socket, unsigned int port, unsigned long non_blocking, int listen_socket);
+ZED_NET_DEF int zed_net_tcp_socket_open(zed_net_socket_t *socket, unsigned short port, unsigned long non_blocking, int listen_socket);
 
 // Connect to a remote endpoint
 // Returns 0 on success.
@@ -264,7 +263,7 @@ ZED_NET_DEF const char *zed_net_host_to_str(unsigned int host) {
     return inet_ntoa(in);
 }
 
-ZED_NET_DEF int zed_net_udp_socket_open(zed_net_socket_t *sock, unsigned int port, unsigned long non_blocking) {
+ZED_NET_DEF int zed_net_udp_socket_open(zed_net_socket_t *sock, unsigned short port, unsigned long non_blocking) {
     if (!sock)
         return zed_net__error("Socket is NULL");
 
@@ -306,7 +305,7 @@ ZED_NET_DEF int zed_net_udp_socket_open(zed_net_socket_t *sock, unsigned int por
     return 0;
 }
 
-ZED_NET_DEF int zed_net_tcp_socket_open(zed_net_socket_t *sock, unsigned int port, unsigned long non_blocking, int listen_socket) {
+ZED_NET_DEF int zed_net_tcp_socket_open(zed_net_socket_t *sock, unsigned short port, unsigned long non_blocking, int listen_socket) {
     if (!sock)
         return zed_net__error("Socket is NULL");
